Allow function bind calls
The example added to the "valid" test was failing because of the `.bind`
call. That exposed that we were expecting all member expressions to
follow our chaining rules, which isn't strictly needed. For now we
simply do not apply the rule to member expressions made to the end of a
function expression. That does not completely fix the problem, but at
least stops the false positives in the provided test case.