Fix comma-dangle to require comma in multiline arrays and objects
In our node standards, we require a trailing comma in multiline
array and object declarations. This fix is to lint for this
condition.

Also, since browsers can blow up with the trailing comma, we
never put it in arrays and objects by default, only in node.

Also fixes missing commas in index.js