Make dangling commas always required, even in default config
Previously the dangling commas (on multi-line objects and arrays) were only required in
the node ruleset. This was because we were worried about them causing a problem in browser
environments, so did not want them there. However, since we always run uglified versions
of our code in the browser, and uglify removes the superfluous trailing commas, we can
enforce our preferred standard on all codebases, including browser codebases.