Add configuration for TypeScript and allow ES6+ features
To support linting TypeScript with ESLint, we need to use a special
parser to tell ESLint how to parse the TypeScript files. We also need to
override some rules that have problems with TypeScripts special syntax.

In addition, in order to use ES6+ features in our code, we need to
enable the correct parser options and allow some previously banned
syntax.

This commit addresses these 2 concerns by adding the TypeScript parser
and adjusting our index.js rules to allow ES6+.