Add editorconfig file
In order to have uniform editor settings across developer machines for
all our projects, we've decided to supply a standard .editorconfig file.
The most logical place to do this seems to be in our linting
configuration, since the settings mostly affect our linting. We can then
symlink the file to the root of each of our projects.

This commit adds the .editorconfig file.