Remove silvermine/indent rule and use base eslint indent rule
A new option on the base eslint indent rule fixes the problem that
forced us to make a custom rule. We can now line up variable
declarations by using the "VariableDeclarator": "first" option.

To use this rule, we need to ensure a peer dependency of
eslint >= 5.11.1. Also, `grunt-eslint` needed to be re-installed to
update it's eslint dependency in the package-lock.json file (the same
version was used, but the eslint dependency pulled down the new version)