feat: adjust how configurations are exposed for flexibility
Now that ESLint favors a flat config, which is essentially just an array of
JS objects, consumers will be able to import just the parts of a config
that they need. For convenience, exposing complete and discrete configs
makes it easier to simply adopt a config when no customization is needed,
or import just the parts that are neccessary when a good deal of
additional configuration will be done by the consuming project.