feat: upgrade to ESLint 9, move silvermine-plugin here
For several years, we've maintained a separate plugin for our custom ESLint rules
(@silvermine/eslint-plugin-silvermine). The plugin itself needs to be linted, so it
required a dependency on this package. This package needed to be linted, so it required a
dependency on a previous version of itself, and the plugin. This circular dependency makes
upgrading ESLint a multi-step process. It's now possible to ship the plugin along with
this package, simplifying the maintenance of both repos.

This commit also updates us to Node 24, and converts the codebase to TypeScript.

BREAKING CHANGE: ESM is now required. Deprecates @silvermine/eslint-plugin-silvermine